infrastructure/pyproject.toml
[tool.poetry]
name = "infrastructure"
version = "1.13.1"
description = "infrastructure as code for diffusion endpoint"
authors = ["Duncan Haywood"]
license = "Prop"

[tool.poetry.dependencies]
python = "^3.8"
aws-cdk-lib = "^2.45.0"
boto3 = "^1.26.12"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
